PythonAPI/requirements.txt
future==0.18.2
gym==0.15.4                                            
jsonpickle==1.4.1
matplotlib==3.1.3
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
numpy==1.18.1
overpy==0.4
pylint==2.4.4
scipy==1.4.1
torch==1.5.0
torchvision==0.6.0
tornado==4.5.3
utm==0.5.0
tensorboardX==2.1.0